# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    Aircraft Engine MRO Decision Support API. Ingests engine condition data,
    reasons over it with a domain ontology, knowledge graph and rules engine,
    and produces traceable MRO Work Recommendations for human planners and for
    SAP S/4HANA Cloud (mock mode).
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engines/{esn}:
    get:
      operationId: getEngine
      tags: [engines]
      summary: Get one engine
      parameters:
        - $ref: "#/components/parameters/EsnParam"
      responses:
        "200":
          description: Engine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engine"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /readings:
    get:
      operationId: getEngineReadings
      tags: [engines]
      summary: Get engine sensor readings (time series)
      parameters:
        - name: engineId
          in: query
          required: true
          schema:
            type: string
          description: Engine serial number (ESN)
        - name: parameter
          in: query
          required: false
          schema:
            type: string
          description: Filter by ECTM parameter code (e.g. EGT_MARGIN)
      responses:
        "200":
          description: Readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParameterReading"
...
  /graph:
    get:
      operationId: getGraph
      tags: [graph]
      summary: Get knowledge-graph nodes and edges
      parameters:
        - name: engineId
          in: query
          required: false
          schema:
            type: string
          description: Scope the graph to one engine's subgraph
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Filter nodes by class/type
      responses:
        "200":
          description: Graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
  /graph/nodes/{id}:
    get:
      operationId: getGraphNode
      tags: [graph]
      summary: Get one graph node
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphNode"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGraphNode
      tags: [graph]
      summary: Correct an instance node (validated against the ontology)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GraphNodeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphNode"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /backtest/runs/{id}:
    get:
      operationId: getBacktestRun
      tags: [backtest]
      summary: Get a backtest run with per-cycle detail
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRun"
        "404":
          $ref: "#/components/responses/NotFound"
  /backtest/run:
    post:
      operationId: runBacktest
      tags: [backtest]
      summary: Replay historical data through the pipeline and evaluate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "200":
          description: Completed run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRun"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    EsnParam:
      name: esn
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    EngineStatus:
      type: string
      enum: [operational, monitor, action_required, grounded]
...
    Engine:
      type: object
      required:
        - esn
        - model
        - tailNumber
        - status
        - healthScore
        - tsn
        - csn
        - tso
        - cso
        - openRecommendations
        - lastUpdated
      properties:
        esn: { type: string }
        model: { type: string }
        tailNumber: { type: string }
        operator: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/EngineStatus" }
        healthScore: { type: integer }
        tsn: { type: number, description: "Time since new (hours)" }
        csn: { type: integer, description: "Cycles since new" }
        tso: { type: number, description: "Time since overhaul (hours)" }
        cso: { type: integer, description: "Cycles since overhaul" }
        egtMargin: { type: ["number", "null"], description: "Current EGT margin (degC)" }
        openRecommendations: { type: integer }
        lastUpdated: { type: string }
    ParameterReading:
      type: object
      required: [engineId, parameter, value, unit, cycle, timestamp, status]
      properties:
        engineId: { type: string }
        parameter: { type: string }
        label: { type: string }
        value: { type: number }
        unit: { type: string }
        cycle: { type: integer }
        timestamp: { type: string }
        status:
          type: string
          enum: [normal, caution, warning]
...
    GraphNode:
      type: object
      required: [id, type, label, properties]
      properties:
        id: { type: string }
        type: { type: string, description: "Ontology class id" }
        label: { type: string }
        properties:
          type: object
          additionalProperties: true
    GraphEdge:
      type: object
      required: [id, source, target, label]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        label: { type: string }
    GraphData:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
    GraphNodeUpdate:
      type: object
      required: [properties]
      properties:
        properties:
          type: object
          additionalProperties: true
...
    BacktestInput:
      type: object
      required: [engineId]
      properties:
        engineId: { type: string }
        name: { type: string }
    BacktestResultRow:
      type: object
      required: [cycle, predicted, actual, outcome]
      properties:
        cycle: { type: integer }
        predicted: { type: boolean }
        actual: { type: boolean }
        outcome:
          type: string
          enum: [true_positive, false_positive, true_negative, false_negative]
        note: { type: ["string", "null"] }
    BacktestRun:
      type: object
      required:
        - id
        - name
        - engineId
        - failureMode
        - totalCycles
        - leadTimeCycles
        - precision
        - recall
        - createdAt
        - rows
      properties:
        id: { type: string }
        name: { type: string }
        engineId: { type: string }
        failureMode: { type: string }
        totalCycles: { type: integer }
        recommendationsGenerated: { type: integer }
        recordedActions: { type: integer }
        leadTimeCycles: { type: integer, description: "Cycles of early warning before recorded action" }
        precision: { type: number }
        recall: { type: number }
        createdAt: { type: string }
        rows:
          type: array
          items:
            $ref: "#/components/schemas/BacktestResultRow"