# lib/api-zod/src/generated/types/engine.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aircraft Engine MRO Decision Support API. Ingests engine condition data, reasons over it with a domain ontology, knowledge graph and rules engine, and produces traceable MRO Work Recommendations for human planners and for SAP S/4HANA Cloud (mock mode).
 * OpenAPI spec version: 0.1.0
 */
import type { EngineStatus } from './engineStatus';

export interface Engine {
  esn: string;
  model: string;
  tailNumber: string;
  /** @nullable */
  operator?: string | null;
  status: EngineStatus;
  healthScore: number;
  /** Time since new (hours) */
  tsn: number;
  /** Cycles since new */
  csn: number;
  /** Time since overhaul (hours) */
  tso: number;
  /** Cycles since overhaul */
  cso: number;
  /**
     * Current EGT margin (degC)
     * @nullable
     */
  egtMargin?: number | null;
  openRecommendations: number;
  lastUpdated: string;
}